requirements.txt
Django==3.1
django-cors-headers==3.5.0
djangorestframework==3.11.1
mysqlclient==2.0.1
django-crispy-forms==1.11.1
